hw5/requirements.txt
simpy>=4.0.1
gym>=0.21.0
numpy>=1.23.0
matplotlib>=3.5.0
pyyaml>=6.0
torch>=1.11.0
stable-baselines3[extra]>=1.7.0
gymnasium
shimmy>=2.0
matplotlib
pandas
tensorboard
ipykernel